% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_setup.R
\name{automatic_setup}
\alias{automatic_setup}
\title{Automaticaly sets up authorization of Twitter API access}
\usage{
automatic_setup(cons_key, cons_secret, acc_token, acc_secret, app_name)
}
\arguments{
\item{cons_key}{Your consumer key}

\item{cons_secret}{Your consumer secret key}

\item{acc_token}{Your consumer access token}

\item{acc_secret}{Your consumer access secret token}

\item{app_name}{The name of your application}
}
\value{
Connects to the API and stores your credentials in a variable "twitter_token" that you can use after to retrieve networks
}
\description{
Use this function prior to get_followers_network to connect to the API. You will need yours API keys and the name of your APP
}
