% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_list.R
\name{write_dataset_list}
\alias{write_dataset_list}
\title{Write dataset list}
\usage{
write_dataset_list(dataframes_list)
}
\arguments{
\item{dataframes_list}{The result of the get_followers_network function. A variable consisting in a list of two dataframes : one for the nodes of the network and the other for the edges}
}
\value{
The csv files are exported in the working directory and can be then used in a network analysing software like Gephi
}
\description{
Exports the results of the get_followers_network function under 2 csv file in your working directory (!! May not yet work with windows)
}
\examples{

\dontrun{

## Get a network of the following list :
network <- get_followers_network(users_list, token = twitter_token, max_accounts = 100000)
# Export the network to your computer and remove the first column (rownames)
write_dataset_list<-(network)

}

}
