\name{auto_setup}
\alias{autosetup}
\title{Automaticly setup the connection to the API}
\usage{
automatic_setup(
    cons_key,
    cons_secret,
    acc_token,
    acc_secret,
    app_name
)
}
\description{
Will connect to the API.
}
\examples{
auto_setup(cons_key,cons_secret,acc_token,acc_secret,app_name)
}
