% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followers_network.R
\name{get_followers_network}
\alias{get_followers_network}
\title{Twitter followers network}
\usage{
get_followers_network(x, token = NULL, max.accounts = 50000)
}
\arguments{
\item{x}{A list of screen names from whom the followers and following will be retreive}

\item{token}{A variable containing the twitter API credentials. If you used automatic_setup, you can put "twitter_token"}

\item{max.accounts}{A maximum number of followers/following. A list of 100 accounts can bring to a million followers/following to retreive for the network. Default is set to 50.000, which is roughly 1 hour long}
}
\value{
A dataframe with 2 columns : an Source node and Target node. This dataframe can be exported in csv and used in a network analysing software like Gephi
}
\description{
Create a network of Twitter followers from a list of accounts
}
